+++ spring-cloud-starter-tencent-polaris-discovery/src/test/java/com/tencent/cloud/polaris/endpoint/PolarisDiscoveryEndpointTest.java
@@ -16,6 +16,7 @@
  */
 package com.tencent.cloud.polaris.endpoint;
 
+import java.util.Collections;
 import java.util.Map;
 
 import com.tencent.cloud.polaris.PolarisDiscoveryProperties;
@@ -38,6 +39,8 @@ import static com.tencent.polaris.test.common.Consts.NAMESPACE_TEST;
 import static com.tencent.polaris.test.common.Consts.PORT;
 import static com.tencent.polaris.test.common.Consts.SERVICE_PROVIDER;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.mock;
 
 /**
  * Test for polaris discovery endpoint.
@@ -75,14 +78,21 @@ public class PolarisDiscoveryEndpointTest {
 	@Test
 	public void testPolarisDiscoveryEndpoint() {
 		this.contextRunner.run(context -> {
-			PolarisDiscoveryProperties polarisDiscoveryProperties = context.getBean(PolarisDiscoveryProperties.class);
-			DiscoveryClient discoveryClient = context.getBean(PolarisDiscoveryClient.class);
+			PolarisDiscoveryProperties polarisDiscoveryProperties = context
+					.getBean(PolarisDiscoveryProperties.class);
+			DiscoveryClient discoveryClient = context
+					.getBean(PolarisDiscoveryClient.class);
+
 			PolarisDiscoveryHandler polarisDiscoveryHandler = context.getBean(PolarisDiscoveryHandler.class);
 			PolarisDiscoveryEndpoint polarisDiscoveryEndpoint = new PolarisDiscoveryEndpoint(polarisDiscoveryProperties, discoveryClient, polarisDiscoveryHandler);
-
 			Map<String, Object> mapInfo = polarisDiscoveryEndpoint.polarisDiscovery("java_provider_test");
-
 			assertThat(polarisDiscoveryProperties).isEqualTo(mapInfo.get("PolarisDiscoveryProperties"));
+
+			DiscoveryClient discoveryClient1 = mock(DiscoveryClient.class);
+			doReturn(Collections.singletonList("xx")).when(discoveryClient1).getServices();
+			PolarisDiscoveryEndpoint polarisDiscoveryEndpoint1 = new PolarisDiscoveryEndpoint(polarisDiscoveryProperties, discoveryClient1, polarisDiscoveryHandler);
+			Map<String, Object> mapInfo2 = polarisDiscoveryEndpoint1.polarisDiscovery(null);
+			assertThat(polarisDiscoveryProperties).isEqualTo(mapInfo2.get("PolarisDiscoveryProperties"));
 		});
 	}